Changed the FrameAllocator trait alloc() function to return NonNull<u8> instead of NonNull<[u8]>. The reason for this change is to better suit the GlobalAlloc trait, which returns a *mut u8 and not a NonNull slice. (I believe the original implementation of the FrameAllocator trait returned a NonNull<[u8]> to match the Rust Allocator trait)
Changed the start attribute inside of FrameAllocatorSolution from
NonNull<u8> to NonNull<[u8]> The reason is so that the struct knows
exactly the size of the memory region it manages andn not just have a
pointer to the start of the region. Although the size of the memory
region can be calculated by using the length of the coremap, this
refactor should make the usage more straightforward.